[Feat] 리뷰 작성 시 이벤트를 발행해 Store의 리뷰 개수, 평점 업데이트 처리

+++ src/main/java/com/zts/delivery/review/application/service/dto/StoreReviewInfo.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 @Builder
 public record StoreReviewInfo(
         StoreId storeId,
-        long reviewCount,
+        int reviewCount,
         BigDecimal averageScore
 ) {
     public static StoreReviewInfo of(StoreReview storeReview) {
